Add client and requests modules

+++ raft/src/main/java/com/raft/server/LogEntry.java
@@ -1,15 +1,15 @@
 package com.raft.server;
 
 final public class LogEntry {
-    private final int term;
+    private final long term;
     private final StateChange change;
 
-    public LogEntry(int term, StateChange change) {
+    public LogEntry(long term, StateChange change) {
         this.term = term;
         this.change = change;
     }
 
-    public int getTerm() {
+    public long getTerm() {
         return term;
     }
